tests/requirements.txt
async-timeout==3.0.1
attrs==22.2.0
certifi==2022.12.7
chardet==3.0.4
idna==2.10
iniconfig==2.0.0
multidict==6.0.4
packaging==23.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.11.0
pytest==6.1.2
pytest-order==1.0.0
python-dotenv==0.19.2
requests==2.25.1
toml==0.10.2
typing-extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
